Extract serial reader and parser into aircube_ui.serial_reader
