+++ scripts/aircube_app.py
@@ -8,9 +8,7 @@ __app_name__ = "AirCube"
 
 import collections
 import csv
-import json
 import os
-import re
 import sys
 from datetime import datetime
 
@@ -20,7 +18,7 @@ from PyQt6.QtWidgets import (
     QGroupBox, QStatusBar, QMessageBox, QSpinBox, QSplitter,
     QFrame, QGridLayout
 )
-from PyQt6.QtCore import QTimer, Qt, QThread, pyqtSignal
+from PyQt6.QtCore import QTimer, Qt
 from PyQt6.QtGui import QFont, QIcon, QAction
 
 import matplotlib
@@ -28,78 +26,15 @@ matplotlib.use('QtAgg')
 from matplotlib.backends.backend_qtagg import FigureCanvasQTAgg as FigureCanvas
 from matplotlib.figure import Figure
 
-import serial
 from serial.tools import list_ports
 
-# JSON pattern for parsing sensor data
-JSON_PATTERN = re.compile(r"\{.*\}")
+from aircube_ui.serial_reader import SerialReaderThread
 
 # CSV header compatible with other AirCube scripts
 CSV_HEADER = [
     "timestamp", "ens210_status", "temperature_c", "temperature_f",
     "humidity", "ens16x_status", "etvoc", "eco2", "aqi"
 ]
-
-
-def parse_json_line(line):
-    """Parse a JSON sensor data line into a flat dict."""
-    match = JSON_PATTERN.search(line)
-    if not match:
-        return None
-    try:
-        data = json.loads(match.group(0))
-        return {
-            "timestamp": data.get("timestamp"),
-            "temperature_c": data["ens210"].get("temperature_c"),
-            "temperature_f": data["ens210"].get("temperature_f"),
-            "humidity": data["ens210"].get("humidity"),
-            "ens210_status": data["ens210"].get("status"),
-            "ens16x_status": data["ens16x"].get("status"),
-            "etvoc": data["ens16x"].get("etvoc"),
-            "eco2": data["ens16x"].get("eco2"),
-            "aqi": data["ens16x"].get("aqi"),
-        }
-    except (KeyError, TypeError, json.JSONDecodeError):
-        return None
-
-
-class SerialReaderThread(QThread):
-    """Background thread for reading serial data."""
-    data_received = pyqtSignal(dict)
-    error_occurred = pyqtSignal(str)
-    
-    def __init__(self, port, baud=115200):
-        super().__init__()
-        self.port = port
-        self.baud = baud
-        self.running = False
-        self.serial = None
-    
-    def run(self):
-        try:
-            self.serial = serial.Serial(self.port, self.baud, timeout=0.1)
-            self.running = True
-            while self.running:
-                try:
-                    line = self.serial.readline()
-                    if line:
-                        decoded = line.decode(errors="ignore").strip()
-                        parsed = parse_json_line(decoded)
-                        if parsed:
-                            self.data_received.emit(parsed)
-                except (serial.SerialException, OSError) as e:
-                    if self.running:
-                        self.error_occurred.emit(str(e))
-                    break
-        except serial.SerialException as e:
-            self.error_occurred.emit(str(e))
-        finally:
-            if self.serial and self.serial.is_open:
-                self.serial.close()
-    
-    def stop(self):
-        self.running = False
-        self.wait(2000)
 
 
 class SensorDisplay(QFrame):
